+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.0.3 (TBD)
 
-* TBD
+* GridPoint to coordinate (CLLocationCoordinate2D) method
 
 ## [1.0.2](https://github.com/ngageoint/grid-ios/releases/tag/1.0.2) (09-14-2022)